Change start number of subuid/subgid for osbs-podman user/group
The subuid and subgid starts from 10_000 for osbs-podman, and the count
is pretty large 1_878_948_191, this prevents new users to be created on
hosts with default settings of SUB_UID_MIN, SUB_GID_MIN and etc.

man useradd:

    SUB_UID_MIN (number), SUB_UID_MAX (number), SUB_UID_COUNT (number)

    If /etc/subuid exists, the commands useradd and newusers (unless the
    user already have subordinate user IDs) allocate SUB_UID_COUNT unused
    user IDs from the range SUB_UID_MIN to SUB_UID_MAX for each new user.

    The default values for SUB_UID_MIN, SUB_UID_MAX, SUB_UID_COUNT are
    respectively 100000, 600100000 and 65536.

so with the default values, all IDs between 100_000 and 600100000 have
already been allocated for our osbs-podman user. Changing the start
number to 10_000_000 makes it possible to create ~151 new users on host:

    (10_000_000-100_000)/65536 ~= 151

Signed-off-by: Qixiang Wan <qwan@redhat.com>