Prevent subscription-manager from creating repos
CLOUDBLD-10433

Mount an rhsm.conf into container builds to make sure subscription
manager will not create redhat.repo when the user runs any dnf/yum
commands in their Dockerfile.

The redhat.repo would have been guaranteed empty anyway, but this way it
will not be created at all and will be deleted if it already exists.
This is more consistent with OSBS 1 behavior.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>